backend/requirements.txt
# Backend Python - Email Classifier API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# AI Providers
openai>=1.10.0
google-generativeai>=0.4.0

# File Processing
PyPDF2>=3.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Development
black>=23.12.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0
